requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
python-dotenv
requests
pydantic
google-genai